Further refactoring and code reorganisation.
